feat(useFormState): add revalidateOnChange dependencies array option, refactor options into object (#94)
Replaces the yupOptions argument with an options object including yupOptions and a new option called
revalidateOnChange, which is an array of arbitrary variables. If any of these variables change since
the last validation, the form will revalidate even if none of the values changed. This can be used
to force revalidation when some asynchronous task is done, like checking a connection from
credentials stored in form state.

BREAKING CHANGE: Any forms using the yupOptions argument will need to wrap it in an object with the
property yupOptions instead of passing it directly as a top-level argument.